+++ experiments/common/plot/plot_hidden_vs_visible.py
@@ -2,33 +2,10 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
 import numpy as np
-from typing import List, Optional
-import matplotlib as mpl
+from typing import Optional
 
 from experiments.common.plot.plotting_style import set_rc_params
-
-
-def filter_group(df: pd.DataFrame, target: List) -> pd.DataFrame:
-    key = "group"
-    return filterdf(key=key, df=df, target=target)
-
-
-def filter_std(df: pd.DataFrame, target: List) -> pd.DataFrame:
-    key = "contexts.default_sample_std_percentage"
-    return filterdf(key=key, df=df, target=target)
-
-
-def filter_cf(df: pd.DataFrame, target: List) -> pd.DataFrame:
-    key = "contexts.context_feature_args"
-    return filterdf(key=key, df=df, target=target)
-
-
-def filterdf(key: str, df: pd.DataFrame, target: List) -> pd.DataFrame:
-    exclude = [e for e in df[key].unique() if e not in target]
-    for ex in exclude:
-        ids = df[key] == ex
-        df = df[~ids]
-    return df
+from experiments.common.plot.preprocessing import filter_group, filter_std, filter_cf
 
 
 def adorn_ax(
